+++ build-scripts/build.sh
@@ -4,5 +4,6 @@ set -e
 npm install --force
 python build-scripts/licensechecker/licensechecker.py
 npm run test
+#npm run validate:themes -- default # Validates themes
 npm run build
 cp -r themes output/themes