+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 | Version | Description |
 |---------|-------------|
+| | Adds validation tool for achievement config files. |
 | 2.0.1 | |
 | | Remove achievement close callback. |
 | 2.0.0 | |