+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 | Version | Description |
 |---------|-------------|
+| | Fix Chrome race condition with removed DOM elements triggering blur event with accessible elements.. |
 | | Add level id to stats if present. |
 | | Adds validation tool for achievement config files. |
 | 2.0.1 | |