+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 | Version | Description |
 |---------|-------------|
+| 2.0.2 | |
+| | transientData automatically passed between screens. |
 | | Fix Chrome race condition with removed DOM elements triggering blur event with accessible elements.. |
 | | Add level id to stats if present. |
 | | Adds validation tool for achievement config files. |